model_manager/requirements.txt
azure_storage==0.37.0
Flask==2.0.2
numpy==1.22.2
pymongo==4.1.0
Werkzeug==2.0.3
